+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add logging configuration flags for workload cluster log collection: `--logging-default-namespaces`, `--logging-enable-node-filtering`, `--logging-enable-network-monitoring`, `--logging-include-events-from-namespaces`, and `--logging-exclude-events-from-namespaces`.
 - Add logging configuration section to Helm chart values for controlling log collection behavior.
 
+### Changed
+
+- Change flag parsing from flag default library to `pflag`
+
 ## [0.55.0] - 2026-01-05
 
 ### Changed